+++ htdocs/forgot_password.php
@@ -11,7 +11,8 @@
  * that were distributed with this source code.
  */
 
+use Eventum\Kernel;
+
 require_once __DIR__ . '/../init.php';
 
-$controller = new Eventum\Controller\ForgotPasswordController();
-$controller->run();
+Kernel::handleRequest();